requirements.txt
langchain
langchain_core
langchain-community
langchain-text-splitters
langchain-openai
langchain_huggingface
langchain_chroma
langchain-groq

python-dotenv
ipykernel
pypdf
bs4
pymupdf
chromadb
wikipedia
sentence_transformers
# faiss-cpu
# duckdb
pandas
openai
duckduckgo_search==5.3.1b1
arxiv
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
streamlit
loguru
singleton-decorator